Create U.S. Senate Twitter\X Accounts for author or EngagingWith file
Use this file to find posts by all the accounts by using the author: operator (or similar for your platform) and enclosing the entire query in parentheses. In the same way you can use the engagingWith: (or similar operator if your platform supports it) to find the reposts, comments, and likes of those engaging with these accounts.